Backport PR #873: Fix issue with student dictionaries when assignments have zero points
Fixes #857

This fixes the issue with students not being displayed in the formgrader. The underlying issue was that `Gradebook.student_dicts()` was returning an empty list, even though the students did in fact exist in the database. This is a edgecase that only seems to occur when there are assignments in the database, but they have a combined zero points. A workaround for the issue is to make sure there is at least one assignment with nonzero points.

cc @jcsutherland