Backport PR #1352: Description of "what is nbgrader?"
Following my pattern in writing a "what is jupyterhub?", I've written
a "what is nbgrader?".  This goes into the technical details of what
is happening under the nbgrader hood, which is an important way for
people to self-support.  It's also what I was looking for when I was
first starting out with nbgrader, and had to piece together by reading
...