Look at the correct collection of Exchange classes
In `nbgrader.apps.baseapp.NbGrader.all_configurable_classes`, there was
a subtle problem in the way it was deciding whether to include classes
from `exchange.default` which meant that none were ever included.

The same work to do this check was being done 5 different times in
unique blocks of code, one of which being incorrectly implemented. To
avoid this type of mistake in the future, I added an inline function to
do that check so it's done in a uniform way every time.