Add linkTo util function
Formgrader could be loaded in one of two different contexts: on its own
from a notebook, or inside an iframe a lab. When navigating to a page
outside the formgrader application, this presents an issue in that these
two contexts need different navigation mechanisms.

Previously, the mechanism was decided by looking at the name of the
jupyter-server application. If it was `jupyter-notebook` (notebook<7),
use href. Otherwise, use postMessage. However, it's possible to use the
notebook interface without `jupyter-notebook`, e.g. `jupyter-server`.

Now the mechanism is decided by context clues inside the browser, i.e.
whether formgrader is inside an iframe.

This commit adds a `linkTo` function which is applied to
jQuery elements via `.map` and applies either href or postMessage nav
mechanisms based on whether the current window is the top window.

Additionally, it completely removes the `is_lab` and `is_jlab` checks
as they were relatively naive and add complexity that doesn't need
to be there.