Resolve warnings in SQLAlchemy field defs (#1963)
Most notably, Grade.cell_type was defined as an association_proxy in the
class, but was later overwritten with a column_property. Because the
mapper had already been made aware of the association_proxy, this raised
a warning. Grade.max_score also had a similar problem, with a
column_property overwriting a regular @property.

This commit fixes the warnings by replacing the in-class field
definitions with `field = None`.

Additionally, remove duplicate definitions for `cell_type_gradecell` and
`cell_type_taskcell`.

Fixes #1946